fix: Project drictory typo 수정
