+++ src/main/java/com/dev/museummate/MuseumMateApplication.java
@@ -1,0 +1,13 @@
+package com.dev.museummate;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MuseumMateApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MuseumMateApplication.class, args);
+    }
+
+}
