+++ src/main/java/com/dev/museummate/configuration/EncrypterConfig.java
@@ -1,0 +1,13 @@
+package com.dev.museummate.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+@Configuration
+public class EncrypterConfig {
+    @Bean
+    public BCryptPasswordEncoder encodePwd() {
+        return new BCryptPasswordEncoder();// password를 인코딩 해줄때 쓰기 위함
+    }
+}
