+++ src/main/java/com/dev/museummate/configuration/JpaAuditingConfiguration.java
@@ -1,0 +1,9 @@
+package com.dev.museummate.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfiguration {
+}
