+++ src/main/java/com/dev/museummate/domain/UserRole.java
@@ -1,0 +1,8 @@
+package com.dev.museummate.domain;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public enum UserRole {
+    ROLE_USER,ROLE_ADMIN;
+}
