+++ src/main/java/com/dev/museummate/domain/dto/ErrorResponse.java
@@ -1,0 +1,11 @@
+package com.dev.museummate.domain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class ErrorResponse {
+    private String errorCode;
+    private String message;
+}
