+++ src/main/java/com/dev/museummate/domain/dto/user/UserJoinResponse.java
@@ -1,0 +1,13 @@
+package com.dev.museummate.domain.dto.user;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserJoinResponse {
+    private String userName;
+
+}
