+++ src/main/java/com/dev/museummate/exception/ExceptionManager.java
@@ -1,0 +1,23 @@
+package com.dev.museummate.exception;
+
+import com.dev.museummate.configuration.Response;
+import com.dev.museummate.domain.dto.ErrorResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+
+public class ExceptionManager {
+    @ExceptionHandler(RuntimeException.class)
+    public ResponseEntity<?> runTimeExceptionHandler(RuntimeException e){
+        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                .body("INTERNAL SERVER ERROR");
+    }
+    @ExceptionHandler(AppException.class)
+    public ResponseEntity<?> appExceptionHandler(AppException e){
+        return ResponseEntity.status(e.getErrorCode().getHttpStatus())
+                .body(Response.error(new ErrorResponse(e.getErrorCode().toString(), e.getMessage())));
+    }
+}
