+++ src/main/java/com/dev/museummate/security/SecurityConfiguration.java
@@ -1,0 +1,41 @@
+package com.dev.museummate.security;
+
+import com.dev.museummate.utils.JwtUtils;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+
+@EnableWebSecurity
+@Configuration
+@RequiredArgsConstructor
+public class SecurityConfiguration {
+
+    private final JwtUtils jwtUtils;
+
+    @Value("${jwt.secret}")
+    private String secretKey;
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
+        return httpSecurity
+                .httpBasic().disable()
+                .csrf().disable()
+                .cors().and()
+                .sessionManagement()
+                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+                .and()
+                .authorizeHttpRequests(authorize -> authorize
+                .requestMatchers("/users/join","/users/login").permitAll()
+                .requestMatchers(HttpMethod.POST).authenticated())
+//                .exceptionHandling().authenticationEntryPoint(new CustomAuthenticationEntryPoint())
+//                .and()
+//                .addFilterBefore(new JwtFilter(jwtUtils, secretKey), UsernamePasswordAuthenticationFilter.class)
+                .build();
+    }
+}
