--- src/main/java/com/dev/musiummate/MuseumMateApplication.java
@@ -1,13 +1,0 @@
-package com.dev.musiummate;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MuseumMateApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MuseumMateApplication.class, args);
-    }
-
-}
