--- src/main/java/com/dev/musiummate/configuration/JpaAuditingConfiguration.java
@@ -1,9 +1,0 @@
-package com.dev.musiummate.configuration;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfiguration {
-}
