--- src/main/java/com/dev/musiummate/domain/UserRole.java
@@ -1,8 +1,0 @@
-package com.dev.musiummate.domain;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public enum UserRole {
-    ROLE_USER,ROLE_ADMIN;
-}
