--- src/main/java/com/dev/musiummate/domain/dto/ErrorResponse.java
@@ -1,11 +1,0 @@
-package com.dev.musiummate.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class ErrorResponse {
-    private String errorCode;
-    private String message;
-}
