--- src/main/java/com/dev/musiummate/domain/dto/user/UserJoinResponse.java
@@ -1,13 +1,0 @@
-package com.dev.musiummate.domain.dto.user;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserJoinResponse {
-    private String userName;
-
-}
