--- src/main/java/com/dev/musiummate/domain/entity/ExhibitionEntity.java
@@ -1,33 +1,0 @@
-package com.dev.musiummate.domain.entity;
-
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ExhibitionEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String name;
-    private String startsAt;
-    private String endsAt;
-    private String price;
-    private String ageLimit;
-    private String detailInfo;
-    private String galleryDetail;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "gallery_id")
-    private GalleryEntity gallery;
-
-}
