--- src/main/java/com/dev/musiummate/exception/ErrorCode.java
@@ -1,15 +1,0 @@
-package com.dev.musiummate.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@AllArgsConstructor
-@Getter
-public enum ErrorCode {
-    DUPLICATE_USERNAME(HttpStatus.CONFLICT, "User name conflict"),
-    DUPLICATE_EMAIL(HttpStatus.CONFLICT, "email conflict");
-
-    private HttpStatus httpStatus;
-    private String message;
-}
