--- src/main/java/com/dev/musiummate/exception/ExceptionManager.java
@@ -1,23 +1,0 @@
-package com.dev.musiummate.exception;
-
-import com.dev.musiummate.configuration.Response;
-import com.dev.musiummate.domain.dto.ErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-
-public class ExceptionManager {
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<?> runTimeExceptionHandler(RuntimeException e){
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body("INTERNAL SERVER ERROR");
-    }
-    @ExceptionHandler(AppException.class)
-    public ResponseEntity<?> appExceptionHandler(AppException e){
-        return ResponseEntity.status(e.getErrorCode().getHttpStatus())
-                .body(Response.error(new ErrorResponse(e.getErrorCode().toString(), e.getMessage())));
-    }
-}
