--- src/main/java/com/dev/musiummate/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.dev.musiummate.repository;
-
-import com.dev.musiummate.domain.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<UserEntity,Long> {
-    Optional<UserEntity> findByUserName(String userName);
-    Optional<UserEntity> findByEmail(String email);
-}
