+++ src/test/java/com/dev/museummate/MuseumMateApplicationTests.java
@@ -1,0 +1,13 @@
+package com.dev.museummate;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MuseumMateApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
