--- src/test/java/com/dev/musiummate/MuseumMateApplicationTests.java
@@ -1,13 +1,0 @@
-package com.dev.musiummate;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MuseumMateApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
